refactor: ChatRoom 메서드 분리
- 네트워크 작업을 ChatRoom으로부터 분리하기 위해
  ChatNetworkManager 생성
- ChatRoom 초기화 시 chatNetworkManager를 외부로부터 주입
  주입하기 위한 ChatNetworkManageable 구현
