+++ StreamChat/StreamChat/Model/ChatRoom.swift
@@ -10,23 +10,14 @@ import UIKit
 final class ChatRoom: NSObject {
     
     // MARK: - Properties
-    var urlSession: URLSessionProtocol
-    var streamTask: URLSessionStreamTask?
     var username = ""
+    var chatNetworkManager: ChatNetworkManageable
     
     // MARK: - Methods
     
-    init(urlSession: URLSessionProtocol = URLSession.shared) {
-        self.urlSession = urlSession
-    }
-    
-    func setUpNetwork() {
-        let configuration = URLSessionConfiguration.default
-        configuration.requestCachePolicy = .reloadIgnoringLocalCacheData
-        urlSession = URLSession(configuration: configuration, delegate: self, delegateQueue: nil)
-        self.streamTask = urlSession.streamTask(withHostName: StreamInformation.host, port: StreamInformation.portNumber)
-        streamTask?.resume()
-        read(from: streamTask)
+    init(chatNetworkManager: ChatNetworkManageable) {
+        self.chatNetworkManager = chatNetworkManager
+        chatNetworkManager.setUpNetwork()
     }
     
     func joinChat(username: String) {
@@ -34,7 +25,7 @@ final class ChatRoom: NSObject {
             
             return
         }
-        write(data: data)
+        chatNetworkManager.write(data: data)
     }
     
     func send(_ message: String) {
@@ -42,49 +33,6 @@ final class ChatRoom: NSObject {
             
             return
         }
-        write(data: data)
-    }
-    
-    func closeStream() {
-        urlSession.invalidateAndCancel()
-    }
-    
-    private func read(from streamTask: URLSessionStreamTask?) {
-        streamTask?.readData(ofMinLength: ConnectionConfiguration.minimumReadLength, maxLength: ConnectionConfiguration.maximumReadLength, timeout: ConnectionConfiguration.timeOut) { [weak self] data, _, error in
-            guard let self = self else {
-                
-                return
-            }
-            defer {
-                self.read(from: streamTask)
-            }
-            guard let data = data else {
-                return
-            }
-            
-            // MARK: - Read Log
-            NSLog(String(data: data, encoding: .utf8) ?? "no message could be read")
-            
-            if let readError = error {
-                NSLog(readError.localizedDescription)
-            }
-        }
-    }
-    
-    private func write(data: Data) {
-        streamTask?.write(data, timeout: ConnectionConfiguration.timeOut) { error in
-            if let writeError = error {
-                NSLog(writeError.localizedDescription)
-            }
-        }
-    }
-}
-
-// MARK: - URLSessionStreamDelegate
-
-extension ChatRoom: URLSessionStreamDelegate {
-    func urlSession(_ session: URLSession, didBecomeInvalidWithError error: Error?) {
-        streamTask?.closeRead()
-        streamTask?.closeWrite()
+        chatNetworkManager.write(data: data)
     }
 }
